Added a way to remove callbacks in connector components.
