+++ slider/src/core/network/messageHandler.cpp
@@ -4,10 +4,6 @@
 
 using namespace Core;
 
-InvokerBase::InvokerBase(const char* name)
-    : name(name)
-{}
-
 MessageHandler::QueueItem::QueueItem()
     :data(0), length(0)
 {}
